@@ -4,21 +4,16 @@
   "private": true,
   "homepage": "https://keithbillings.com/",
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@popperjs/core": "^2.6.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "aos": "^2.3.4",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-github-calendar": "^2.0.2",
-    "react-hook-form": "^6.15.4",
-    "react-material-ui-carousel": "^2.2.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.1",
-    "react-tooltip": "^4.2.15"
+    "@testing-library/jest-dom": "^5.15.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.8.0",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "^5.0.1",
+    "react-tooltip": "^5.11.1",
+    "sass": "^1.62.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -44,6 +39,6 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.1.0"
+    "gh-pages": "^5.0.0"
   }
 }
