@@ -1,11 +1,10 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
+import ReactDOM from "react-dom/client";
+
+// Components
 import App from "./App";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById("root")
-);
+// Styles
+import "./index.css";
+
+ReactDOM.createRoot(document.getElementById("root")).render(<App />);
